fixed sizing and file names
