+++ src/components/banner/wrenches.js
@@ -7,15 +7,14 @@ import { withStyles } from '@material-ui/core/styles';
 
 const styles = theme => ({
 	bigWrench: {
-		width: '53vw',
+		width: '450px',
 		left: '5%',
 		[theme.breakpoints.down('xs')]: {
 			display: 'none'
 		}
 	},
 	smallWrench: {
-		width: '21vw',
-		marginLeft: '7vw',
+		width: '120px',
 		display: 'none',
 		[theme.breakpoints.down('xs')]: {
 			display: 'block'
