+++ src/pages/playground.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import HomePage from '../components/homepage/homepagetemplate';
+import Banner from '../components/banner/banner';
 
 import HeadFooter from '../components/headfooter/headfooter';
 export default class PlayGround extends React.Component {
 	render() {
 		return (
 			<HeadFooter>
-				<HomePage/>
+				<Banner/>
 			</HeadFooter>
 		);
 	}
